requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cytoolz==0.12.0
db-dtypes==1.0.5
eth-abi==2.2.0
eth-account==0.5.9
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-tester==0.6.0b7
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.7.1
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
hexbytes==0.2.3
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.3
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.23.5
oauthlib==3.2.2
packaging==21.3
pandas==1.5.2
pandas-gbq==0.18.1
parsimonious==0.8.1
proto-plus==1.22.1
protobuf==3.19.5
py-ecc==5.2.0
py-evm==0.5.0a3
py-geth==3.10.0
py-solc-x==1.1.1
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.16.0
pydata-google-auth==1.4.0
pyethash==0.1.27
pyparsing==3.0.9
pyrsistent==0.19.2
pysha3==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rlp==2.0.1
rsa==4.9
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
toolz==0.12.0
tqdm==4.64.1
trie==2.0.0a5
typing-extensions==3.10.0.2
uniswap-python==0.6.2
urllib3==1.26.13
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
